Attempt at a Loading Screen, Refiement of Special Attacks + HitBox and Damaging Boss Work
Long day... computer was using a lot of resources and slowing the frame rate of the game down in the first 10 or so seconds of running it. Tried implementing a loading screen, but problem was solved by just exiting some other programs.

Also did a lot of work setting up hitboxes and damage scripts for special attacks. Slashes needed own scripting as it is just an extension of the player's sprite, not a new game object being instantiated.

Lots of little bugs piling up, need to do a refactor and bug killing-spree soon.