Layers, Tags and Prefabs Fixed, Hitboxes and AttackBoxes Working
Start of a restructuring with reorganizing and defining of layers and tags and how they interact with each other. In turn made possible to get player/enemy attack hitboxes only be able to interact with the other's hazard boxes (damage taking zone) and not full sprite zone. This allows the player to hit the boss easier without getting hit first. Especially important for melee attacks.